working user and order routes EXCEPT userId/orders

+++ server/db/models/order.js
@@ -2,10 +2,10 @@ const Sequelize = require('sequelize')
 const db = require('../db')
 
 const Order = db.define('order', {
-  subtotal: Sequelize.INTEGER,
-  total: Sequelize.INTEGER,
-  shippingCost: Sequelize.INTEGER,
-  tax: Sequelize.INTEGER,
+  subtotal: Sequelize.DECIMAL(10,2),
+  total: Sequelize.DECIMAL(10,2),
+  shippingCost: Sequelize.DECIMAL(10,2),
+  tax: Sequelize.FLOAT,
   shippingAddress: Sequelize.STRING,
   billingAddress: Sequelize.STRING,
   creditCardNumber: Sequelize.BIGINT,
